// src/app/tokens/interfaces/tv-result.interface.ts
export interface TvResult {
  media_type: string;
  adult: boolean;
  backdrop_path: string;
  genre_ids: number[];
  id: number;
  original_language: string;
  original_name: string;
  overview: string;
  popularity: number;
  poster_path: string;
  first_air_date: string;
  name: string;
  origin_country: string[];
  vote_average: number;
  vote_count: number;
}